requirements.txt
aiogram==3.22.0
asyncio==4.0.0
Django==5.1.1
djangorestframework==3.15.2
flake8==7.1.1
pyTelegramBotAPI==4.22.1
python-dotenv==1.0.1
requests==2.32.3